requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
numpy==1.23.4
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.5
scipy==1.9.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
